# postman/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new product",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all products",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get one product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "summary": "Get one product",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update one product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one product",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete one product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one product",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/product-variants": {
      "get": {
        "operationId": "ProductsController_getAllProductVariantsOfProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all product-variants of this product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductVariantEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all product-variants of this product",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}/reviews": {
      "get": {
        "operationId": "ProductsController_getAllReviewsOfProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all reviews of this product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all reviews of this product",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEntity"
                }
              }
            }
          }
        },
        "summary": "Create a new review",
        "tags": [
          "Reviews"
        ]
      },
      "get": {
        "operationId": "ReviewsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all reviews",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "get": {
        "operationId": "ReviewsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEntity"
                }
              }
            }
          }
        },
        "summary": "Get a review",
        "tags": [
          "Reviews"
        ]
      },
      "patch": {
        "operationId": "ReviewsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEntity"
                }
              }
            }
          }
        },
        "summary": "Update a review",
        "tags": [
          "Reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewEntity"
                }
              }
            }
          }
        },
        "summary": "Delete a review",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}/media-list": {
      "get": {
        "operationId": "ReviewsController_getAllMediaOfReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all media of review",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all media of review",
        "tags": [
          "Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "BE Ecomerce Shop",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "MediaEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "url": {
            "type": "string",
            "example": "https://s3.amazonaws.com/bucket/file.jpg"
          },
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "DOCUMENT"
            ],
            "example": "IMAGE"
          },
          "reviewId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "userId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "productVariantId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "isShopLogo": {
            "type": "boolean",
            "example": false
          },
          "isShopBanner": {
            "type": "boolean",
            "example": false
          },
          "isCategoryFile": {
            "type": "boolean",
            "example": false
          },
          "isAvatarFile": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "type",
          "reviewId",
          "userId",
          "productVariantId",
          "isShopLogo",
          "isShopBanner",
          "isCategoryFile",
          "isAvatarFile",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProductEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Product Name"
          },
          "description": {
            "type": "object",
            "example": "Product description",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 99.99
          },
          "stockKeepingUnit": {
            "type": "string",
            "example": "SKU123456"
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "categoryId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "voucherId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "shopOfficeId": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "stockKeepingUnit",
          "stock",
          "categoryId",
          "createByUserId",
          "voucherId",
          "shopOfficeId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductVariantEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "productId": {
            "format": "int64",
            "type": "integer",
            "example": 1315
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 852
          },
          "variantName": {
            "type": "string",
            "example": "name of product variant"
          },
          "variantColor": {
            "type": "string",
            "example": "red"
          },
          "variantSize": {
            "type": "string",
            "example": "XL"
          },
          "price": {
            "type": "number",
            "example": 46546
          },
          "stock": {
            "type": "number",
            "example": 851
          },
          "stockKeepingUnit": {
            "type": "string",
            "example": "EWDGDSED715545D"
          },
          "voucherId": {
            "type": "object",
            "example": 1325,
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "productId",
          "createByUserId",
          "variantName",
          "variantColor",
          "variantSize",
          "price",
          "stock",
          "stockKeepingUnit",
          "voucherId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReviewEntity": {
        "type": "object",
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "productId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "userId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "productVariantId": {
            "format": "int64",
            "type": "integer",
            "example": 1
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "object",
            "example": "Great product!",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "productId",
          "userId",
          "productVariantId",
          "rating",
          "comment",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sample product name"
          },
          "description": {
            "type": "string",
            "example": "Sample product description"
          },
          "price": {
            "type": "number",
            "example": 12
          },
          "stockKeepingUnit": {
            "type": "string",
            "example": "ADSFDSAF1463218FA"
          },
          "stock": {
            "type": "number",
            "example": 25
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 1231
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:18.484Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:18.484Z"
          },
          "categoryId": {
            "format": "int64",
            "type": "integer",
            "example": 1325
          },
          "voucherId": {
            "format": "int64",
            "type": "integer",
            "example": 1325
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "stockKeepingUnit",
          "stock",
          "createByUserId",
          "createdAt",
          "updatedAt",
          "categoryId",
          "voucherId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sample product name"
          },
          "description": {
            "type": "string",
            "example": "Sample product description"
          },
          "price": {
            "type": "number",
            "example": 12
          },
          "stockKeepingUnit": {
            "type": "string",
            "example": "ADSFDSAF1463218FA"
          },
          "stock": {
            "type": "number",
            "example": 25
          },
          "createByUserId": {
            "format": "int64",
            "type": "integer",
            "example": 1231
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:18.484Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:18.484Z"
          },
          "categoryId": {
            "format": "int64",
            "type": "integer",
            "example": 1325
          },
          "voucherId": {
            "format": "int64",
            "type": "integer",
            "example": 1325
          }
        }
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "productId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "userId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "productVariantId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "comment": {
            "type": "string",
            "example": "Perfect"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:20.189Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:20.189Z"
          }
        },
        "required": [
          "productId",
          "userId",
          "productVariantId",
          "rating",
          "comment",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {
          "productId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "userId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "productVariantId": {
            "format": "int64",
            "type": "integer",
            "example": 851
          },
          "rating": {
            "type": "number",
            "example": 5
          },
          "comment": {
            "type": "string",
            "example": "Perfect"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:20.189Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-12-04T09:24:20.189Z"
          }
        }
      },
...
    }
  }
}